+++ LMS-Learning Management System/Models/CardSubject.cs
@@ -13,7 +13,7 @@ namespace LMS_Learning_Management_System.Models
         public int ClassId { get; set; }
         public int Semester { get; set; }
         public string TeacherId { get; set; }
-        public double? PaymentAmount { get; set; }
+        public decimal? PaymentAmount { get; set; }
         public bool? IsPayment { get; set; }
         public DateTime? PaymentDate { get; set; }
         public virtual Card CardNoNavigation { get; set; }
